# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Knowledge Base API
  description: |
    知识库管理系统API文档
    
    ## 认证方式
    
    ### 业务API认证
    使用API Key进行认证，在请求头中添加：
    ```
    Authorization: Bearer kb_live_xxxx
    ```
    
    ### 管理后台认证
    使用JWT Token进行认证，先通过登录接口获取token，然后在请求头中添加：
    ```
    Authorization: Bearer eyJhbGciOiJIUzI1NiIs...
    ```
    
    ## 权限说明
    - `read`: 读取权限
    - `write`: 写入权限
    - `admin`: 管理权限
    - `super_admin`: 超级管理员权限
    
  version: 1.0.0
  contact:
    email: support@example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000/api/v1
    description: 本地开发环境
  - url: http://limousea.asia:3000/api/v1
    description: 生产环境
...
paths:
...
  /categories:
    get:
      tags:
        - Categories
      summary: 获取分类列表
      description: 返回所有分类及其关联的知识条目数量
      security:
        - ApiKeyAuth: []
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /admin/categories:
    get:
      tags:
        - Admin Categories
      summary: 获取分类列表
      description: 返回所有分类及其关联的知识条目数量，支持分页和搜索
      security:
        - AdminAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: search
          in: query
          description: 按分类名称模糊搜索
          schema:
            type: string
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminCategoryListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
    
    post:
      tags:
        - Admin Categories
      summary: 创建分类
      security:
        - AdminAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCategoryRequest'
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          description: 分类名称已存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /admin/categories/{id}:
    get:
      tags:
        - Admin Categories
      summary: 获取分类详情
      security:
        - AdminAuth: []
      parameters:
        - $ref: '#/components/parameters/CategoryId'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    
    put:
      tags:
        - Admin Categories
      summary: 更新分类
      security:
        - AdminAuth: []
      parameters:
        - $ref: '#/components/parameters/CategoryId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCategoryRequest'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: 分类名称已存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    
    delete:
      tags:
        - Admin Categories
      summary: 删除分类
      description: 只能删除没有关联知识条目的分类
      security:
        - AdminAuth: []
      parameters:
        - $ref: '#/components/parameters/CategoryId'
      responses:
        '204':
          description: 删除成功
        '400':
          description: 分类下存在知识条目，无法删除
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /admin/categories/{id}/stats:
    get:
      tags:
        - Admin Categories
      summary: 获取分类统计信息
      description: 返回指定分类的详细统计，包括按语言、难度、框架分布，平均质量分数，最近新增条目数等
      security:
        - AdminAuth: []
      parameters:
        - $ref: '#/components/parameters/CategoryId'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminCategoryStats'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  # ==================== 知识条目管理 ====================
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: http
      scheme: bearer
      bearerFormat: API Key
      description: API Key认证，格式为 kb_live_xxxx
    
    AdminAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 管理员JWT Token认证
  parameters:
...
    CategoryId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: 分类ID
  responses:
    BadRequest:
      description: 请求参数错误
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: 未授权
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            timestamp:
              type: string
              format: date-time
...
    Category:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          maxLength: 100
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CategoryWithEntryCount:
      type: object
      properties:
        name:
          type: string
        entry_count:
          type: integer
          description: 该分类下的知识条目数量
    CategoryListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/CategoryWithEntryCount'
        total:
          type: integer
          description: 分类总数
    AdminCategoryListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            type: object
            allOf:
              - $ref: '#/components/schemas/Category'
              - type: object
                properties:
                  entry_count:
                    type: integer
                    description: 该分类下的知识条目数量
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    CreateCategoryRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
    UpdateCategoryRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
...
    AdminCategoryStats:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        total_entries:
          type: integer
          description: 该分类下的知识条目总数
        by_language:
          type: object
          additionalProperties:
            type: integer
          description: 按编程语言统计的条目数量
        by_difficulty:
          type: object
          additionalProperties:
            type: integer
          description: 按难度等级(1-5)统计的条目数量
        by_framework:
          type: object
          additionalProperties:
            type: integer
          description: 按框架统计的条目数量
        avg_quality_score:
          type: number
          format: float
          description: 平均质量分数
        recent_entries:
          type: integer
          description: 最近7天新增的条目数量

    # ========== 认证相关 ==========
...